youtube-chatbot-backend/requirements.txt
youtube-transcript-api
fastapi
uvicorn
python-dotenv
langchain-community 
langchain-google-genai 
faiss-cpu 
tiktoken 
python-dotenv 
google-ai-generativelanguage